Impl. autofarm functionality
created AutoMiner.cs to handle autofarm capabilities; added AutoMiner objects to scene for each currency type
updated CurrencyObject and BuildingObject to store respective amounts and purchase costs as floats (but still display in scene as ints)

+++ Assets/Scripts/Game/HUD/BuildingShop.cs
@@ -34,7 +34,7 @@ public class BuildingShop : MonoBehaviour
     public void PurchaseBuilding(BuildingObject buildingObject)
     {
         CurrencyObject currency = currencyObjects[(int)buildingObject.costCurrencyType];
-        int purchaseCost = buildingObject.purchaseCost;
+        float purchaseCost = buildingObject.purchaseCost;
 
         if (currency.amount >= purchaseCost)
         {
