+++ Assets/Scripts/Scriptable Objects/UserData.cs
@@ -5,7 +5,7 @@ using UnityEngine;
 [CreateAssetMenu(fileName = "New User Data", menuName = "Scriptable Object/User Data")]
 public class UserData : ScriptableObject
 {
-    public SerializableDictionary<CurrencyType, int> currencies;
+    public SerializableDictionary<CurrencyType, float> currencies;
     public SerializableDictionary<BuildingType, int> buildings;
 
     public int buildingTypesUnlocked;
